# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local personal memory server: episodic (OpenSearch) + knowledge (Neo4j) derived from canonical hot JSON, archived to S3.",
        "title": "memory-mcp v2 API",
        "contact": {},
        "version": "2.0"
    },
    "host": "127.0.0.1:8420",
    "basePath": "/",
    "paths": {
...
        "/v1/documents/{sha}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Download original document bytes by sha256",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sha256",
                        "name": "sha",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/documents/{sha}/chunks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "List document_chunk episodes for a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sha256",
                        "name": "sha",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/episode.Record"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
...
        "/v1/{ws}/{team}/{proj}/documents": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Ingest a document (multipart) — blob + chunks + document node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "document file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/document.IngestResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "503": {
                        "description": "document pipeline unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierr.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "document.IngestResult": {
            "type": "object",
            "properties": {
                "blob_key": {
                    "description": "BlobKey is the S3 key the original was uploaded to (cold-first).",
                    "type": "string"
                },
                "chunk_ids": {
                    "description": "ChunkIDs are the created document_chunk episode ids, in sequence order.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "degraded": {
                    "description": "Degraded names every derived plane whose best-effort mirror failed. The\nhot write succeeded regardless (§1), so an ingest that could not index\nits chunks or MERGE its document node still reports 201 — but it says so\nhere rather than claiming a mirror that does not exist (§0 principle 3).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extractable": {
                    "description": "Extractable is false when no deterministic text layer exists (e.g.\nscanned PDF); the agent handles such documents itself.",
                    "type": "boolean"
                },
                "node_id": {
                    "description": "NodeID is the auto-created knowledge document node id.",
                    "type": "string"
                },
                "sha": {
                    "description": "SHA is the lowercase hex sha256 of the original bytes.",
                    "type": "string"
                },
                "truncated": {
                    "description": "Truncated is non-nil when the chunk cap cut the tail.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/document.Truncation"
                        }
                    ]
                }
            }
        },
        "document.Truncation": {
            "type": "object",
            "properties": {
                "indexed": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "episode.Actor": {
            "type": "string",
            "enum": [
                "agent",
                "user",
                "system"
            ],
            "x-enum-varnames": [
                "ActorAgent",
                "ActorUser",
                "ActorSystem"
            ]
        },
...
        "episode.Kind": {
            "type": "string",
            "enum": [
                "event",
                "conversation",
                "decision",
                "observation",
                "document_chunk"
            ],
            "x-enum-varnames": [
                "KindEvent",
                "KindConversation",
                "KindDecision",
                "KindObservation",
                "KindDocumentChunk"
            ]
        },
        "episode.Record": {
            "type": "object",
            "properties": {
                "actor": {
                    "description": "Actor is agent|user|system.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Actor"
                        }
                    ]
                },
                "consolidated": {
                    "description": "Consolidated marks the record as distilled into knowledge; a\nprecondition for cold archival (§3). Never inferred by the server: it is\nset only where the agent states the distillation itself, by naming this\nrecord in the provenance of a POST .../knowledge/nodes (§3, §0\nprinciple 2).",
                    "type": "boolean"
                },
                "entities": {
                    "description": "Entities are normalized entity names mentioned by this record.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "ID is a ULID; immutable forever (provenance anchor).",
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is one of the values accepted by ValidKind.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Kind"
                        }
                    ]
                },
                "last_recalled": {
                    "type": "string"
                },
                "occurred_at": {
                    "description": "OccurredAt is the event time (drives TTL aging, §3.1).",
                    "type": "string"
                },
                "recall_count": {
                    "description": "RecallCount / LastRecalled track search hits. LastRecalled is RFC3339\nor \"\" when never recalled.",
                    "type": "integer"
                },
                "refs": {
                    "description": "Refs is set only for kind=document_chunk.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/episode.Refs"
                        }
                    ]
                },
                "text": {
                    "description": "Text is the body — the nori full-text indexing target.",
                    "type": "string"
                }
            }
        },
        "episode.Refs": {
            "type": "object",
            "properties": {
                "chunk_seq": {
                    "type": "integer"
                },
                "doc_sha": {
                    "type": "string"
                }
            }
        },
...
        "httpserver.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apierr.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}